backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

# AI/ML dependencies
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.7.0
transformers==4.40.0
torch==2.1.1
faiss-cpu==1.7.4
pypdf==4.2.0
python-docx==1.1.0
unstructured==0.12.4
tiktoken==0.6.0

# Vector database
pgvector==0.2.4

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
numpy==1.25.2
pandas==2.1.4

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0